[JS]Gather stamps tests. Extract tests of checking first stamp to separated file.
